Moved the new Resource middleware to a new folder
